set checkout_only troughout the app

+++ app/controllers/cart_controller_decorator.rb
@@ -2,7 +2,7 @@
 
 module CartControllerDecorator
   def show
-    apply_sparta_discount(spree_current_order, spree_current_user)
+    apply_sparta_discount(spree_current_order, spree_current_user, check_only = true)
 
     super
   end
@@ -19,7 +19,7 @@ module CartControllerDecorator
       private_metadata: add_item_params[:private_metadata],
       options: add_item_params[:options]
     )
-    apply_sparta_discount(spree_current_order, spree_current_user)
+    apply_sparta_discount(spree_current_order, spree_current_user, check_only = true)
     spree_current_order.reload
 
     render_order(result)
@@ -31,7 +31,7 @@ module CartControllerDecorator
     spree_authorize! :update, spree_current_order, order_token
 
     result = set_item_quantity_service.call(order: spree_current_order, line_item: line_item, quantity: params[:quantity])
-    apply_sparta_discount(spree_current_order, spree_current_user)
+    apply_sparta_discount(spree_current_order, spree_current_user, check_only = true)
 
     render_order(result)
   end
@@ -50,7 +50,7 @@ module CartControllerDecorator
     result = associate_service.call(guest_order: guest_order, user: spree_current_user)
 
     if result.success?
-      apply_sparta_discount(guest_order, spree_current_user)
+      apply_sparta_discount(guest_order, spree_current_user, check_only = true)
       guest_order.reload
       render_serialized_payload { serialize_resource(guest_order) }
     else
@@ -65,26 +65,28 @@ module CartControllerDecorator
   end
 
   def remove_coupon_code
-    apply_sparta_discount(spree_current_order, spree_current_user)
+    apply_sparta_discount(spree_current_order, spree_current_user, check_only = true)
 
     super
   end
 
   private
 
-  def apply_sparta_discount(order, user)
+  def apply_sparta_discount(order, user, check_only)
     return unless order.line_items.any? && user.present?
     return unless user.public_metadata["spl_no_card"].present?
 
-    check_only = true
     spl_response = SpartaLoyaltyService.send_request(order.number,
                                                      user.public_metadata["spl_no_card"],
                                                      order.line_items,
                                                      DateTime.current,
                                                      order.products,
                                                      check_only)
-    spree_current_order.promotions.delete_all if spree_current_order.promotions.any?
-    create_sparta_adjustments(spl_response, order)
+
+    if spl_response.present?
+      spree_current_order.promotions.delete_all if spree_current_order.promotions.any?
+      create_sparta_adjustments(spl_response, order)
+    end
   end
 
   def create_sparta_adjustments(spl_response, order)
